+++ src/rofi_rbw/selector.py
@@ -1,16 +1,16 @@
 import re
 from subprocess import run
-from typing import List, Tuple, Union
+from typing import Dict, List, Tuple, Union
 
-from .abstractionhelper import is_wayland, is_installed
+from .abstractionhelper import is_installed, is_wayland
 from .credentials import Credentials
 from .entry import Entry
-from .models import Action, Target, Targets, CANCEL, DEFAULT, SYNC
+from .models import Action, Keybinding, Target, Targets
 
 
 class Selector:
     @staticmethod
-    def best_option(name: str = None) -> 'Selector':
+    def best_option(name: str = None) -> "Selector":
         try:
             return next(selector for selector in Selector.__subclasses__() if selector.name() == name)()
         except StopIteration:
@@ -32,28 +32,30 @@ class Selector:
         entries: List[Entry],
         prompt: str,
         show_help_message: bool,
-        additional_args: List[str]
-    ) -> Tuple[Union[List[Target], DEFAULT, CANCEL], Union[Action, DEFAULT, CANCEL, SYNC], Union[Entry, None]]:
+        keybindings: Dict[str, Tuple[Action, List[Target]]],
+        additional_args: List[str],
+    ) -> Tuple[Union[List[Target], None], Union[Action, None], Union[Entry, None]]:
         raise NoSelectorFoundException()
 
     def select_target(
         self,
         credentials: Credentials,
         show_help_message: bool,
-        additional_args: List[str]
-    ) -> Tuple[Union[List[Target], CANCEL], Union[Action, DEFAULT, CANCEL]]:
+        keybindings: Dict[str, Action],
+        additional_args: List[str],
+    ) -> Tuple[Union[List[Target], None], Union[Action, None]]:
         raise NoSelectorFoundException()
 
     def _format_targets_from_credential(self, credentials: Credentials) -> List[str]:
         targets = []
         if credentials.username:
-            targets.append(f'Username: {credentials.username}')
+            targets.append(f"Username: {credentials.username}")
         if credentials.password:
             targets.append(f'Password: {credentials.password[0]}{"*" * (len(credentials.password) - 1)}')
         if credentials.has_totp:
             targets.append(f'TOTP: {credentials.totp}')
         if len(credentials.uris) == 1:
-            targets.append(f'URI: {credentials.uris[0]}')
+            targets.append(f"URI: {credentials.uris[0]}")
         else:
             for (key, value) in enumerate(credentials.uris):
                 targets.append(f'URI {key + 1}: {value}')
@@ -63,13 +65,13 @@ class Selector:
         return targets
 
     def _extract_targets(self, output: str) -> List[Target]:
-        return [Target(line.split(':')[0]) for line in output.strip().split('\n')]
+        return [Target(line.split(":")[0]) for line in output.strip().split("\n")]
 
 
 class Rofi(Selector):
     @staticmethod
     def supported() -> bool:
-        return is_installed('rofi')
+        return is_installed("rofi")
 
     @staticmethod
     def name() -> str:
@@ -80,8 +82,9 @@ class Rofi(Selector):
         entries: List[Entry],
         prompt: str,
         show_help_message: bool,
-        additional_args: List[str]
-    ) -> Tuple[Union[List[Target], DEFAULT, CANCEL], Union[Action, DEFAULT, CANCEL, SYNC], Union[Entry, None]]:
+        keybindings: List[Keybinding],
+        additional_args: List[str],
+    ) -> Tuple[Union[List[Target], None], Union[Action, None], Union[Entry, None]]:
         parameters = [
             'rofi',
             '-markup-rows',
@@ -90,55 +93,29 @@ class Rofi(Selector):
             '-sort',
             '-p',
             prompt,
-            '-kb-custom-11',
-            'Alt+p',
-            '-kb-custom-12',
-            'Alt+u',
-            '-kb-custom-13',
-            'Alt+t',
-            '-kb-custom-14',
-            'Alt+m',
-            '-kb-custom-15',
-            'Alt+s',
-            *additional_args
+            *self.__build_parameters_for_keybindings(keybindings),
+            *additional_args,
         ]
 
-        if show_help_message:
-            parameters.extend([
-                '-mesg',
-                '<b>Alt+1</b>: Autotype username and password | <b>Alt+2</b> Type username | <b>Alt+3</b> Type password | <b>Alt+u</b> Copy username | <b>Alt+p</b> Copy password | <b>Alt+t</b> Copy totp | <b>Alt+s</b> Sync'
-            ])
+        if show_help_message and keybindings:
+            parameters.extend(self.__format_keybindings_message(keybindings))
 
-        rofi = run(parameters, input="\n".join(self.__format_entries(entries)), capture_output=True, encoding="utf-8")
+        rofi = run(
+            parameters,
+            input='\n'.join(self.__format_entries(entries)),
+            capture_output=True,
+            encoding='utf-8'
+        )
 
         if rofi.returncode == 1:
-            return CANCEL(), CANCEL(), None
-        elif rofi.returncode == 24:
-            return DEFAULT(), SYNC(), None
-        elif rofi.returncode == 10:
-            return_action = Action.TYPE
-            return_targets = [Targets.USERNAME, Targets.PASSWORD]
-        elif rofi.returncode == 11:
-            return_action = Action.TYPE
-            return_targets = [Targets.USERNAME]
-        elif rofi.returncode == 12:
-            return_action = Action.TYPE
-            return_targets = [Targets.PASSWORD]
-        elif rofi.returncode == 20:
-            return_action = Action.COPY
-            return_targets = [Targets.PASSWORD]
-        elif rofi.returncode == 21:
-            return_action = Action.COPY
-            return_targets = [Targets.USERNAME]
-        elif rofi.returncode == 22:
-            return_action = Action.COPY
-            return_targets = [Targets.TOTP]
-        elif rofi.returncode == 23:
-            return_action = DEFAULT
-            return_targets = [Targets.MENU]
+            return None, Action.CANCEL, None
+        elif rofi.returncode >= 10:
+            keybinding = keybindings[(rofi.returncode - 10)]
+            return_action = keybinding.action
+            return_targets = keybinding.targets
         else:
-            return_action = DEFAULT()
-            return_targets = DEFAULT()
+            return_action = None
+            return_targets = None
 
         return return_targets, return_action, self.__parse_formatted_string(rofi.stdout)
 
@@ -155,27 +132,22 @@ class Rofi(Selector):
         self,
         credentials: Credentials,
         show_help_message: bool,
-        additional_args: List[str]
-    ) -> Tuple[Union[List[Target], CANCEL], Union[Action, DEFAULT, CANCEL]]:
+        keybindings: List[Keybinding],
+        additional_args: List[str],
+    ) -> Tuple[Union[List[Target], None], Union[Action, None]]:
         parameters = [
-            'rofi',
-            '-markup-rows',
-            '-dmenu',
-            '-p',
+            "rofi",
+            "-markup-rows",
+            "-dmenu",
+            "-p",
             'Choose target',
-            '-i',
-            '-kb-custom-11',
-            'Alt+t',
-            '-kb-custom-12',
-            'Alt+c',
-            *additional_args
+            "-i",
+            *self.__build_parameters_for_keybindings(keybindings),
+            *additional_args,
         ]
 
-        if show_help_message:
-            parameters.extend([
-                '-mesg',
-                '<b>Alt+t</b>: Type | <b>Alt+c</b> Copy'
-            ])
+        if show_help_message and keybindings:
+            parameters.extend(self.__format_keybindings_message(keybindings))
 
         rofi = run(
             parameters,
@@ -185,15 +157,41 @@ class Rofi(Selector):
         )
 
         if rofi.returncode == 1:
-            return CANCEL(), CANCEL()
-        elif rofi.returncode == 20:
-            action = Action.TYPE
-        elif rofi.returncode == 21:
-            action = Action.COPY
+            return None, Action.CANCEL
+        elif rofi.returncode >= 10:
+            action = keybindings[(rofi.returncode - 10)].action
         else:
-            action = DEFAULT()
+            action = None
 
         return (self._extract_targets(rofi.stdout)), action
+
+    def __build_parameters_for_keybindings(self, keybindings: List[Keybinding]) -> List[str]:
+        params = []
+        for index, keybinding in enumerate(keybindings):
+            params.extend([f"-kb-custom-{1 + index}", keybinding.shortcut])
+
+        return params
+
+    def __format_keybindings_message(self, keybindings: List[Keybinding]):
+        return [
+            "-mesg",
+            " | ".join(
+                [
+                    f"<b>{keybinding.shortcut}</b>: {self.__format_action_and_targets(keybinding)}"
+                    for keybinding in keybindings
+                ]
+            ),
+        ]
+
+    def __format_action_and_targets(self, keybinding: Keybinding) -> str:
+        if keybinding.targets and Targets.MENU in keybinding.targets:
+            return "Menu"
+        elif keybinding.action == Action.SYNC:
+            return "Sync logins"
+        elif keybinding.targets:
+            return f"{keybinding.action.value.title()} {', '.join([target.raw for target in keybinding.targets])}"
+        else:
+            return keybinding.action.value.title()
 
 
 class Wofi(Selector):
@@ -210,21 +208,16 @@ class Wofi(Selector):
         entries: List[Entry],
         prompt: str,
         show_help_message: bool,
-        additional_args: List[str]
-    ) -> Tuple[Union[List[Target], DEFAULT, CANCEL], Union[Action, DEFAULT, CANCEL, SYNC], Union[Entry, None]]:
-        parameters = [
-            'wofi',
-            '--dmenu',
-            '-p',
-            prompt,
-            *additional_args
-        ]
+        keybindings: Dict[str, Tuple[Action, List[Target]]],
+        additional_args: List[str],
+    ) -> Tuple[Union[List[Target], None], Union[Action, None], Union[Entry, None]]:
+        parameters = ["wofi", "--dmenu", "-p", prompt, *additional_args]
 
         wofi = run(parameters, input="\n".join(self.__format_entries(entries)), capture_output=True, encoding="utf-8")
         if wofi.returncode == 0:
-            return DEFAULT(), DEFAULT(), self.__parse_formatted_string(wofi.stdout)
+            return None, None, self.__parse_formatted_string(wofi.stdout)
         else:
-            return CANCEL(), CANCEL(), None
+            return None, Action.CANCEL, None
 
     def __format_entries(self, entries: List[Entry]) -> List[str]:
         max_width = max(len(it) for it in entries)
@@ -239,27 +232,22 @@ class Wofi(Selector):
         self,
         credentials: Credentials,
         show_help_message: bool,
-        additional_args: List[str]
-    ) -> Tuple[Union[List[Target], CANCEL], Union[Action, DEFAULT, CANCEL]]:
-        parameters = [
-            'wofi',
-            '--dmenu',
-            '-p',
-            'Choose target',
-            *additional_args
-        ]
+        keybindings: Dict[str, Action],
+        additional_args: List[str],
+    ) -> Tuple[Union[List[Target], None], Union[Action, None]]:
+        parameters = ["wofi", "--dmenu",'-p', 'Choose target', *additional_args]
 
         wofi = run(
             parameters,
-            input='\n'.join(self._format_targets_from_credential(credentials)),
+            input="\n".join(self._format_targets_from_credential(credentials)),
             capture_output=True,
-            encoding='utf-8'
+            encoding="utf-8",
         )
 
         if wofi.returncode == 1:
-            return CANCEL(), CANCEL()
+            return None, Action.CANCEL
 
-        return self._extract_targets(wofi.stdout), DEFAULT()
+        return self._extract_targets(wofi.stdout), None
 
 
 class NoSelectorFoundException(Exception):